refactor: 集中配置到JacksonConfig
