+++ agileboot-infrastructure/src/main/java/com/agileboot/infrastructure/config/ApplicationConfig.java
@@ -1,9 +1,6 @@
 package com.agileboot.infrastructure.config;
 
-import java.util.TimeZone;
 import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 
@@ -18,11 +15,4 @@ import org.springframework.context.annotation.EnableAspectJAutoProxy;
 @MapperScan("com.agileboot.orm.**.mapper")
 public class ApplicationConfig {
 
-    /**
-     * 时区配置
-     */
-    @Bean
-    public Jackson2ObjectMapperBuilderCustomizer jacksonObjectMapperCustomization() {
-        return jacksonObjectMapperBuilder -> jacksonObjectMapperBuilder.timeZone(TimeZone.getDefault());
-    }
 }
