+++ agileboot-infrastructure/src/main/java/com/agileboot/infrastructure/config/JacksonConfig.java
@@ -1,5 +1,7 @@
-package com.agileboot.infrastructure.security.xss;
+package com.agileboot.infrastructure.config;
 
+import com.agileboot.infrastructure.security.xss.JsonHtmlXssTrimSerializer;
+import java.util.TimeZone;
 import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
@@ -8,7 +10,7 @@ import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
  * @author valarchie
  */
 @Configuration
-public class JsonXssConfiguration implements Jackson2ObjectMapperBuilderCustomizer{
+public class JacksonConfig implements Jackson2ObjectMapperBuilderCustomizer{
 
 //    这种配置方式会覆盖 yml中的jackson配置， 使用下面的customize配置则不会
 //    @Bean
@@ -20,7 +22,10 @@ public class JsonXssConfiguration implements Jackson2ObjectMapperBuilderCustomiz
 
     @Override
     public void customize(Jackson2ObjectMapperBuilder jacksonObjectMapperBuilder) {
+        // 防XSS脚本注入
         jacksonObjectMapperBuilder.deserializers(new JsonHtmlXssTrimSerializer());
+        // 默认时区配置
+        jacksonObjectMapperBuilder.timeZone(TimeZone.getDefault());
     }
 
 }
